Fixed http cookie issue
